getgifs helper
